Revert "Test renaming JS file"
This reverts commit 1c1b7399dd46be47ab1e8492d65fb7fbabdeb1a9.
