+++ src/libs/ComposerFocusManager.js
